requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
typer>=0.9.0
pydantic>=2.0.0
browser-use>=0.1.0
langchain-openai>=0.0.5 